Package m4atool using setuptools.
