+++ m4atool/__init__.py
@@ -1,0 +1,1 @@
+from .m4atool import M4a
